+++ src/pulumi/core-dtap/analytics/databricks/workspaces/__init__.py
@@ -1,7 +1,0 @@
-# Init the platform outputs
-from project_config import platform_outputs
-
-platform_outputs["analytics"]["databricks"]["workspaces"] = {}
-
-from . import engineering
-from . import analytics